% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_prefix.R
\name{add_prefix}
\alias{add_prefix}
\title{Add a prefix in a dataframe column}
\usage{
add_prefix(prefix, data, column)
}
\arguments{
\item{prefix}{specify a prefix for the wanted columns in data.}

\item{data}{specify data frame}

\item{column}{specify a dataframe containing variables}
}
\value{
returns a dataframe with prefix added
}
\description{
This function used to add a prefix in a dataframe column or columns
}
\examples{
add_prefix("G_", head(iris, 2))
}
