% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_col_miss.R
\name{check_col_miss}
\alias{check_col_miss}
\title{Chech for columns missing values}
\usage{
check_col_miss(data, decreasing = TRUE, transpos = NULL)
}
\arguments{
\item{data}{specify dataframe}

\item{decreasing}{TRUE/FALSE order missing values}

\item{transpos}{TRUE/FALSE}
}
\value{
returns a dataframe containing number of missing values per dataframe column
}
\description{
This function test for missing values in a dataframe columns
}
