% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_row_miss.R
\name{check_row_miss}
\alias{check_row_miss}
\title{Chech for rows missing values}
\usage{
check_row_miss(data, decreasing = NULL, transpos = NULL)
}
\arguments{
\item{data}{specify dataframe}

\item{decreasing}{TRUE/FALSE order missing values}

\item{transpos}{TRUE/FALSE}
}
\value{
returns a dataframe containing number of missing values per dataframe row
}
\description{
This function test for missing values in a dataframe rows
}
