% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr_string.R
\name{chr_string}
\alias{chr_string}
\title{Return character into a vector (you can copy into r source file instead of writing it manually)}
\usage{
chr_string(x, writeline = NULL)
}
\arguments{
\item{x}{specify character or string vector to be converted}

\item{writeline}{TRUE/FALSE whether to return interpretable written language or not}
}
\value{
returns a dataframe
}
\description{
This function returns characters into a vector (you can copy into r source file instead of writing it manually)
}
\examples{
\dontrun{
letters[1:4]
chr_string(letters[1:4])
chr_string(letters[1:4], writeline = T)
chr_string(letters[1:4], writeline = F)
}
}
