% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_row_col.R
\name{del_cols}
\alias{del_cols}
\title{Delete columns in a dataframe}
\usage{
del_cols(c, df)
}
\arguments{
\item{c}{specify columns to be deleted}

\item{df}{specify dataframe}
}
\value{
returns a dataframe
}
\description{
This function delete columns in a dataframe
}
\examples{
\dontrun{
del_cols(c(1, 3), data)
del_cols(c("1", "3"), data)
}
}
