% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del_row_col.R
\name{del_rows}
\alias{del_rows}
\title{Delete rows in a dataframe}
\usage{
del_rows(r, df)
}
\arguments{
\item{r}{specify rows to be deleted}

\item{df}{specify dataframe}
}
\value{
returns a dataframe
}
\description{
This function delete rows in a dataframe
}
\examples{
\dontrun{
del_rows(c(1, 3), data)
del_rows(c("1", "3"), data)
}
}
