% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop-columns.R
\name{drop_columns}
\alias{drop_columns}
\title{Drops specified columns from a given data set}
\usage{
drop_columns(data, columns)
}
\arguments{
\item{data}{A data frame that includes the columns to be dropped}

\item{columns}{A character vector of column names to drop}
}
\value{
A new data frame with the specified columns dropped
}
\description{
Drops specified columns from a given data set
}
\examples{
drop_columns(data.frame(a = c(1, 2, 3),
                        b = c(4, 5, 6),
                        c = c(7, 8, 9)), c("a", "c"))
drop_columns(data.frame(a = c(1, 2, 3),
                        b = c(4, 5, 6),
                        c = c(7, 8, 9)), "b")

# If columns vector is null, the original data frame is returned unchanged,
# an error is thrown if a column name in columns vector does not exist in given data.

}
