% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_plots.R
\name{facet_plots}
\alias{facet_plots}
\title{Automatically create faceted plot with ggplot2}
\usage{
facet_plots(variable, group, title = "Title", dat)
}
\arguments{
\item{variable}{specify columns}

\item{group}{specify a variable to be faceted with}

\item{title}{title of the plot}

\item{dat}{specify dataframe}
}
\value{
returns a faceted plot
}
\description{
This function Automatically create faceted plot with ggplot2
}
