% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categ_var.R
\name{get_categ_var}
\alias{get_categ_var}
\title{Get categorical variable in a dataframe}
\usage{
get_categ_var(data)
}
\arguments{
\item{data}{specify a dataframe containing variables}
}
\value{
returns a dataframe with frequancy table
}
\description{
get categorical variable in a dataframe
}
