% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_var.R
\name{get_num_var}
\alias{get_num_var}
\title{Get numerical variable in a dataframe}
\usage{
get_num_var(data)
}
\arguments{
\item{data}{specify a dataframe containing variables}
}
\value{
returns a dataframe with frequancy table
}
\description{
get numerical variable in a dataframe
}
