% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_name.R
\name{get_package_name}
\alias{get_package_name}
\title{Get package name}
\usage{
get_package_name(pkg)
}
\arguments{
\item{pkg}{specify columns to be deleted}
}
\value{
returns package name
}
\description{
Get package name
}
\examples{
\dontrun{
get_package_name("wrangler")
}
}
