% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getpercent.R
\name{getpercent}
\alias{getpercent}
\title{Get percentage of a "variable" in data}
\usage{
getpercent(data, data1, var)
}
\arguments{
\item{data}{specify dataframe}

\item{data1}{specify second dataframe}

\item{var}{specify data frame}
}
\value{
returns a dataframe
}
\description{
get percentage of a "variable" in data, which is separated by another variable, from the the total total "variable" in data1
}
