% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_col.R
\name{index_col}
\alias{index_col}
\title{Get index of a dataframe column by name, or get a name by index}
\usage{
index_col(data, column = NULL, id = NULL)

index_col(data, column = NULL, id = NULL)
}
\arguments{
\item{data}{must be a dataframe}

\item{column}{specify the column name to return the index}

\item{id}{specify the index of wanted column to return the name}
}
\value{
the index of column name, or the column name of the index

the index of column name, or the column name of the index
}
\description{
Get index of a dataframe column by name, or get a name by index

Get index of a dataframe column by name, or get a name by index
}
\details{
Provide either an id or a name to return a name or an id repectively

Provide either an id or a name to return a name or an id repectively
}
\examples{
\dontrun{
index_col(iris, column = "Sepal.Length")
index_col(iris, id = 1)
index_col(iris, id = c(1, 2))
}
\dontrun{
index_col(iris, column = "Sepal.Length")
index_col(iris, id = 1)
index_col(iris, id = c(1, 2))
}
}
\author{
Ibrahim H. Elkhidir, originally

Ibrahim H. Elkhidir, originally
}
