% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_val.R
\name{index_val}
\alias{index_val}
\title{Get index of a value in dataframe, row or colum or both}
\usage{
index_val(data, value, ind = NULL, pair = NULL)
}
\arguments{
\item{data}{specify a dataframe}

\item{value}{specify the of which the index will be returned}

\item{ind}{specify the index to be returned: "row", "col","both"}

\item{pair}{TRUE/FALSE if the index to be returned as pairs like [x,y]}
}
\value{
the index of a value: within a column, a row, or both like [x,y]
}
\description{
Get index of a value in dataframe, row or colum or both
}
\author{
Ibrahim H. Elkhidir
}
