% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.wholenumber.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Test a number for being an integer or not}
\usage{
is.wholenumber(x)
}
\arguments{
\item{x}{a number or a vector of numbers}
}
\description{
This function test a number or a vector of number for being integer or not
}
\examples{
\dontrun{
is.wholenumber(4)
TRUE

is.wholenumber(4.5)
FALSE

is.wholenumber(c(1, 3, 4.5, 6, 7.3))
}
}
