% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_movies_data.R
\name{load_movies_data}
\alias{load_movies_data}
\title{Load movies and myMovieData dataset in global environment}
\usage{
load_movies_data()
}
\description{
Load movies and myMovieData dataset in global environment
}
\examples{
load_movies_data()

}
