% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpackages.R
\name{loadpackages}
\alias{loadpackages}
\title{Loading packages}
\usage{
loadpackages(pkg, ...)
}
\arguments{
\item{pkg}{specifiy packages names}

\item{...}{specify additional packages separated by ",".

For example: "pkg1","pkg2"}
}
\value{
the grid layout, invisibly
}
\description{
Loading packages
}
\examples{
\dontrun{
loadpackages("dplyr", "sjPlot", "wrangler")
}
}
\author{
Ibrahim H. Elkhidir, originally
}
