% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot.R
\name{make_scatter_plot}
\alias{make_scatter_plot}
\title{Scatter plots for budget vs revenue analysis}
\usage{
make_scatter_plot(
  data,
  x,
  y,
  title,
  x_lab,
  y_lab,
  smooth_method = "lm",
  se = TRUE,
  point_alpha = 1,
  pred_line = NULL,
  pred_line_color = "blue"
)
}
\arguments{
\item{data}{A data frame}

\item{x}{column name for the x axis}

\item{y}{column name for the y axis}

\item{title}{Plot title}

\item{x_lab}{x-axis label}

\item{y_lab}{y-axis label}

\item{smooth_method}{Passed to ggplot2::geom_smooth() \code{method}
Use \code{NULL} for no smooth}

\item{se}{Whether to show the smooth confidence ribbon}

\item{point_alpha}{Alpha for ggplot2::geom_point()}

\item{pred_line}{Optional; name of the column of fitted y values}

\item{pred_line_color}{Color for the prediction line}
}
\value{
A ggplot2::ggplot scatterplot
}
\description{
Scatter plots for budget vs revenue analysis
}
