% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mchisq_test.R
\name{mchisq.test}
\alias{mchisq.test}
\title{Get Chi-square test of a specific variable with other categorical variables in a dataframe}
\usage{
mchisq.test(data, var)
}
\arguments{
\item{data}{specify a dataframe containing variables}

\item{var}{specify a categorical variable to get the chi-square tests}
}
\value{
returns a dataframe with chi-square test
}
\description{
Get Chi-square test of a specific variable with other categorical variables in a dataframe
}
