% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrfreqtable.R
\name{mrfreqtable}
\alias{mrfreqtable}
\title{Get multiple response frequency tables}
\usage{
mrfreqtable(data, question.prefix, colname = NULL)
}
\arguments{
\item{data}{specify a dataframe containing variables}

\item{question.prefix}{specify prefix for the the wanted questions in data.}

\item{colname}{specify the name of question variable in the returned freq table.}
}
\value{
returns a dataframe with frequancy table
}
\description{
get percentage of a "variable" in data, which is separated by another variable, from the the total total "variable" in data1
}
