% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearrcolor.R
\name{nearRcolor}
\alias{nearRcolor}
\title{Getting near r colors}
\usage{
nearRcolor(
  rgb,
  cSpace = c("hsv", "rgb255", "Luv", "Lab"),
  dist = switch(cSpace, hsv = 0.1, rgb255 = 30, Luv = 15, Lab = 12)
)
}
\arguments{
\item{rgb}{show the col}

\item{cSpace}{specify the number of rows}

\item{dist}{specify the number of columns}
}
\value{
the grid layout, invisibly
}
\description{
Getting near r colors
}
\author{
Ibrahim H. Elkhidir, originally
}
