% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{not in}
\alias{not in}
\alias{\%!in\%}
\title{\strong{x not in y: \%!in\%}}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{first or left arm}

\item{y}{second or right arm}
}
\description{
This function used to negate the "\code{\%in\%}" operator = \strong{x not in y}
}
\details{
x \code{\%!in\%} y == \strong{x not in y}
}
