% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_names.R
\name{other_names}
\alias{other_names}
\title{get remaining names of a dataframe}
\usage{
other_names(data, name)
}
\arguments{
\item{data}{specify a dataframe object}

\item{name}{specify the column to be be sliced in quotation marks ""}
}
\description{
this function used to get remaining names of a dataframe
}
\examples{
\dontrun{
othernames(iris, "Species")
}
}
