% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCol.R
\name{plotCol}
\alias{plotCol}
\title{Comparing Colors}
\usage{
plotCol(col, nrow = 1, ncol = ceiling(length(col)/nrow), txt.col = "black")
}
\arguments{
\item{col}{give a color or a vector of colors}

\item{nrow}{number of rows for the the color to be displayed in}

\item{ncol}{number of columns for the the color to be displayed in}

\item{txt.col}{default is black}
}
\value{
the grid layout, invisibly
}
\description{
Comparing Colors
}
\author{
Marius Hofert, originally
}
