% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_numbers.R
\name{prime_numbers}
\alias{prime_numbers}
\title{Get the prime numbers of up to a number(n)}
\usage{
prime_numbers(n)
}
\arguments{
\item{n}{integer number}
}
\description{
This function print the prime numbers from lowest up to an integer value(n)
}
