% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reload.R
\name{reload}
\alias{reload}
\title{Reload a package or name space}
\usage{
reload(pkg)
}
\arguments{
\item{pkg}{specify package name}
}
\description{
Reload a package or name space
}
\examples{
reload("dplyr")
}
