% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamer_Old.R
\name{renamer_Old}
\alias{renamer_Old}
\title{Renaming dataframe columns (Old version similar to plyr::rename())}
\usage{
renamer_Old(x, replace, warn_missing = TRUE, warn_duplicated = TRUE)
}
\arguments{
\item{x}{a dataframe}

\item{replace}{original value and replace_with value}

\item{warn_missing}{TRUE/FALSE}

\item{warn_duplicated}{TRUE/FALSE}
}
\description{
a wrapper function for renaming functions in plyr package
}
