% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_col_miss.R
\name{rm_col_miss}
\alias{rm_col_miss}
\title{Remove columns with missing values}
\usage{
rm_col_miss(data)
}
\arguments{
\item{data}{specify a dataframe containing variables}
}
\value{
returns a dataframe with columns with missing values removed
}
\description{
This function remove columns with missing values
}
