% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_miss.R
\name{rm_miss}
\alias{rm_miss}
\title{remove row or columns with percentage of missing values}
\usage{
rm_miss(data, dim, percent_miss)
}
\arguments{
\item{data}{specify dataframe}

\item{dim}{1 for rows, or 2 for columns}

\item{percent_miss}{numeric. cut of number for row or column which is equal or above to be removed}
}
\value{
returns a dataframe containing remaining row or columns
}
\description{
This function test for missing values in a dataframe rows and columns, then removes rows or columns with percentage of missing values
}
