% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_prefix.R
\name{rm_prefix}
\alias{rm_prefix}
\title{Remove a prefix in a dataframe column}
\usage{
rm_prefix(prefix, data, column)
}
\arguments{
\item{prefix}{specify a prefix for the wanted columns in data.}

\item{data}{specify data frame}

\item{column}{specify a dataframe containing variables}
}
\value{
returns a dataframe with prefix removed
}
\description{
This function used to Remove a prefix in a dataframe column or columns
}
\examples{
df <- add_prefix("G_", head(iris, 2))
rm_prefix("G_", df)
}
