% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_row_miss.R
\name{rm_row_miss}
\alias{rm_row_miss}
\title{Remove rows with missing values}
\usage{
rm_row_miss(data)
}
\arguments{
\item{data}{specify a dataframe containing variables}
}
\value{
returns a dataframe with rows with missing values removed
}
\description{
This function remove rows with missing values
}
