% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_suffix.R
\name{rm_suffix}
\alias{rm_suffix}
\title{Remove a suffix in a dataframe column}
\usage{
rm_suffix(suffix, data, column)
}
\arguments{
\item{suffix}{specify a suffix for the wanted columns in data.}

\item{data}{specify data frame}

\item{column}{specify a dataframe containing variables}
}
\value{
returns a dataframe with suffix removed
}
\description{
This function used to Remove a suffix in a dataframe column or columns
}
\examples{
df <- add_suffix("_G", head(iris, 2))
rm_suffix("_G", df)
}
