% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{Generate random sample from a random numbers}
\usage{
sampler(n, N, mean, sd, lower = NULL, upper = NULL, seed = NULL)
}
\arguments{
\item{n}{number of sampled data to be returned from data with size (N), (mean), and (sd)}

\item{N}{Total numbers to be generated}

\item{mean}{mean of generated random numbers (N)}

\item{sd}{standard deviation of generated random numbers (N)}

\item{lower}{minimum value possible in sampled data}

\item{upper}{maximum value possible in sampled data}

\item{seed}{the seed of normal random generation (for reproducibility)}
}
\value{
returns sampled numbers of length (n)
}
\description{
This function generates random data from certain parameters: sample size, total generated pool pf data (N),  mean(N), sd(N), seed, minimum and maximum values
}
\examples{
\dontrun{
sampler(4, 200, 7, 5)
sampler(4, 200, 7, 5, lower = 8)
sampler(4, 200, 7, 5, upper = 8, seed = 4)
}
}
