% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_col.R
\name{slice_col}
\alias{slice_col}
\title{Slice a dataframe column by a seperator}
\usage{
slice_col(
  data,
  split.col,
  sep = ",",
  type = NULL,
  mode = NULL,
  fill = NA,
  drop = FALSE,
  nearcol = TRUE,
  addcol = F,
  fixed = TRUE,
  sort.col = F,
  decreasing = NULL
)
}
\arguments{
\item{data}{specify a dataframe object}

\item{split.col}{specify the column to be be sliced in quotation marks ""}

\item{sep}{specify a separator to slice the column text with (",",";","-",..)}

\item{type}{TRUE/FALSE whether to print the unique column names or not}

\item{mode}{binary/value whether to return values or transform it into 1.}

\item{fill}{specify a value to fill the results NA with}

\item{drop}{whether to drop the column specfified or not}

\item{nearcol}{whether to put splitted col near the original col}

\item{addcol}{whether to paste colname on splitted columns}

\item{fixed}{allowing for regular expression}

\item{sort.col}{whether to sort the the resulting sliced columns or not}

\item{decreasing}{if sorted, whether to be in descending or ascending}
}
\description{
this function used to slice a column with a separator
}
\author{
Ibrahim H. Elkhidir
}
