% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_col.R
\name{slice_col_old}
\alias{slice_col_old}
\title{Slice a dataframe column by a seperator}
\usage{
slice_col_old(
  data,
  column,
  sep,
  return_names = NULL,
  trans = NULL,
  fill = NULL,
  replace = FALSE
)
}
\arguments{
\item{data}{specify a dataframe object}

\item{column}{specify the column to be be sliced in quotation marks ""}

\item{sep}{specify a separator to slice the column text with (",",";","-",..)}

\item{return_names}{TRUE/FALSE whether to print the unique column names or not}

\item{trans}{TRUE/FALSE whether to return values or trans form it into 1.}

\item{fill}{specify a value to fill the results NA with}

\item{replace}{whether to replace or not}
}
\description{
this function used to slice a column with a separator
}
\author{
Ibrahim H. Elkhidir
}
