% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_calc.R
\name{stat_calc}
\alias{stat_calc}
\title{Calculate summary statistic for numerical variables}
\usage{
stat_calc(data, decimal = NULL)
}
\arguments{
\item{data}{specify a dataframe}

\item{decimal}{round the values to a number of decimal}
}
\description{
Calculate summary statistic for numerical variables
}
\examples{
stat_calc(iris[, 1:4])
stat_calc(iris[, 1:4], 3)
}
