% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryse.R
\name{summarySE}
\alias{summarySE}
\title{Automatically create summary standard error for plotting error bar in ggplot2}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = TRUE,
  conf.interval = 0.95,
  drop = TRUE
)
}
\arguments{
\item{data}{specify dataframe}

\item{measurevar}{specify measure variables for mean, n, sd, se, ci}

\item{groupvars}{specify grouping variables}

\item{na.rm}{TRUE/FALSE whether to remove missing values or not}

\item{conf.interval}{specify confidence interval}

\item{drop}{TRUE/FALSE specify whether to drop or not}
}
\value{
returns a dataframe with n, mean, sd, ci, se
}
\description{
This function creates summary standard error for plotting error bar in ggplot2
}
\examples{
\dontrun{
dsummary <- summarySE(iris,
  measurevar = "Petal.Width",
  groupvars = c("Species", "Sepal.Length")
)
dsummary <- summarySE(iris,
  measurevar = "Sepal.Length",
  groupvars = c("Species", "Petal.Width"), F
)
ggplot(dsummary, aes(Sepal.Length, Petal.Width)) +
  geom_point() +
  geom_errorbar(aes(ymin = Petal.Width - se, ymax = Petal.Width + se), width = .03) +
  geom_smooth()
}
}
